+++ test/constant.ts
@@ -31,4 +31,6 @@ export const EXPECTED_MESSAGE_SIGNATURE =
   "0x727c2e31ae342588b680dfc502f0d6a7b8d0f8b9afc4ca313bdad9dca80429741f50b78c2c98ac2f18c4ec1e8fade88c8d7477766d6ceeb1f3a3ddbe7d80e90f1c";
 export const EXPECTED_LONG_TYPED_DATA_SIGNATURE =
   "0x26b713388823e1b9f91cb0064bdaa0a41b3728d35fd0f6bcfdd65900d27d20a21e3d207ded1a97f928c0f3eebdd709a950b353323f32080dad7db4fc8ff8222a1c";
+export const EXPECTED_SHORT_TYPED_DATA_SIGNATURE =
+  "0x66996a481e09815a4400be0988a28c53b357584b15f8dba6804050b8706d5b26645695436dca610fdb9a3ad052a2c35d890e60e8ed7f25beec531743b1aa583e1b";
 export const ACCOUNT_ADDRESS = "0x50707153077cFf1A48192311A12a5f905976AF14";
