+++ test/dapp/index.html
@@ -12,8 +12,8 @@
 <body>
     <!-- <button class="connect-button">Connect</button> -->
     <!-- <button class="sign-button">Sign</button> -->
-    <button class="sign-typedData-button">Sign Typed Data</button>
-    <button class="sign-short-typedData-button">Sign Short Typed Data</button>
+    <!-- <button class="sign-typedData-button">Sign Typed Data</button> -->
+    <!-- <button class="sign-short-typedData-button">Sign Short Typed Data</button> -->
     <button class="increase-button">Increase</button>
     <button class="increase-fees-button">Increase Fees</button>
     <button class="transfer-button">Transfer</button>
