+++ test/testPageFunctions/index.ts
@@ -1,3 +1,4 @@
 export * from "./sign";
 export * from "./requestAccounts";
 export * from "./signLongTypedData";
+export * from "./signShortTypedData";
